feat: Enhance teacher results page and related UI components
- Updated TResultsPage to improve UI elements, including color adjustments and layout refinements.
- Added pull-to-refresh functionality to refresh results in TResultsPage.
- Introduced loading and error states for better user feedback during data fetching.
- Enhanced TeacherCoursePage with a new title section and improved header layout.
- Integrated insights loading in TeacherCoursePage for real-time data validation.
- Updated TeacherEditCategoryNewUI to load and display group data dynamically.
- Improved teacher reports UI with expandable categories and evaluation cards.
- Refactored navigation to use consistent routes for evaluation results.
- Added integration tests for refreshing results and loading states in student and teacher pages.

+++ lib/domain/services/teacher_insights_domain_service.dart
@@ -180,7 +180,7 @@ class TeacherInsightsDomainService {
     return winners;
   }
 
-  bool _isValidScore(int score) =>
+  bool _isValidScore(double score) =>
       score >= minValidScore && score <= maxValidScore;
 
   double _round1(double value) => double.parse(value.toStringAsFixed(1));
@@ -291,7 +291,7 @@ class _StatsBucket {
   final String parentId;
   final String parentName;
 
-  int _sum = 0;
+  double _sum = 0;
   int _count = 0;
 
   _StatsBucket({
@@ -308,7 +308,7 @@ class _StatsBucket {
     return _sum / _count;
   }
 
-  void add(int score) {
+  void add(double score) {
     _sum += score;
     _count++;
   }
